started splitting create_app into separate files
